% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction col_flux.R
\name{col_flux}
\alias{col_flux}
\title{Fonction col_flux}
\usage{
col_flux(data, seq.data, palette = "Paired")
}
\arguments{
\item{data}{}

\item{seq.data}{}
}
\value{
un vecteur avec les différentes valeurs des couleurs et dont le nom des lignes est un level
}
\description{
Cette fonction permet d'avoir une légende universelle pour les graphiques de flux
}
