% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction colId.R
\name{colId}
\alias{colId}
\title{Fonction colId}
\usage{
colId(df)
}
\arguments{
\item{df}{une table data.frame}
}
\value{
les indices de colonnes pouvant être des identifiants
}
\description{
Fonction colId
}
