% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mise en forme tableau cluster.R
\name{Creation_indicateur}
\alias{Creation_indicateur}
\title{Fonction Creation_indicateur}
\usage{
Creation_indicateur(nb_cluster_min, nb_cluster_max, mat_dist, intialclust)
}
\arguments{
\item{nb_cluster_min}{nombre minimal de groupes souhaité}

\item{nb_cluster_max}{nombre maximal de groupes souhaité}

\item{mat_dist}{matrice de distance}

\item{intialclust}{même paramètre que dans la fonction WeightedCluster::wcKMedoids}
}
\value{
un data.frame avec les différents indicateurs de la fonction WeightedCluster::wcKMedoids et le nombre de groupes correspondant
}
\description{
Cette fonction permet de récupérer les différents indicateurs de la fonction WeightedCluster::wcKMedoids et le nombre de groupes correspondant. Il permet de créer le paramètre indicateurs de la fonction tableau_cluster.
}
