% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction data_cluster.R
\name{data_cluster}
\alias{data_cluster}
\title{Title}
\usage{
data_cluster(tabl_ind, data, nb_groupe)
}
\arguments{
\item{tabl_ind}{liste retournée par la fonction "Creation_indicateur"}

\item{data}{jeu de données}

\item{nb_groupe}{Nombre de groupes souhaité}
}
\value{
un data frame avec en dernière colonne les groupes
}
\description{
Title
}
