% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fonction.R
\name{flux_grp}
\alias{flux_grp}
\title{Title}
\usage{
flux_grp(data, seq.data, col_periode, label = NULL, cpal = NULL,
  var_grp, label_grp, box_width = NULL, box_label_cex = NULL,
  arrow_type = "simple", arrow_rez = NULL, title = NULL)
}
\arguments{
\item{data}{}

\item{seq.data}{}

\item{col_periode}{}

\item{label}{}

\item{cpal}{}

\item{var_grp}{}

\item{label_grp}{}

\item{box_width}{}

\item{box_label_cex}{}

\item{arrow_type}{}

\item{arrow_rez}{}

\item{title}{}
}
\description{
Title
}
