% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction graph sous sequences.R
\name{graph_sous_sequences}
\alias{graph_sous_sequences}
\title{Fonction graph_sous_sequences}
\usage{
graph_sous_sequences(subs)
}
\arguments{
\item{subs}{sortie de la fonction seqefsub du package TraMineR}
}
\value{
Un graphique
}
\description{
Fonction graph_sous_sequences
}
