% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervert_dataframes.R
\name{intervert_df}
\alias{intervert_df}
\title{intervert_df()}
\usage{
intervert_df(
  list.output = res.o,
  list.input = res.i,
  equivalences = c(a15_o = "a15_i", a16_o = "a16_i", a22_o = "a22_i")
)
}
\description{
intervert_df() insert dans list.output les data.frames choisies dans list.input selon les equivalences de nom spécifiées
}
\examples{
library(ViCaTraj)
data("EXAMPLE_LIST")
res->res.o
rm(res)
res.i<-lapply(res.o, function(x){mtcars})
names(res.i)<-paste0(names(res.i), "_i")
names(res.o)<-paste0(names(res.o), "_o")
res.3<-intervert_df(list.output=res.o, list.input=res.i, equivalences=c("a15_o"="a15_i", "a16_o"="a16_i", "a22_o"="a22_i"))
}
