% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R, R/module_select_and_plot.R,
%   R/module_select_and_plot2.R
\name{module_data}
\alias{module_data}
\alias{module_select_and_plot}
\alias{module_select_and_plot2}
\title{server function}
\usage{
module_data(input, output, session)

module_select_and_plot(input, output, session, data)

module_select_and_plot2(input, output, session, data)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data.comp}{}
}
\description{
server function

server function

server function
}
