% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_classification.R
\name{module_classification}
\alias{module_classification}
\title{server function}
\usage{
module_classification(input, output, session, data)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data}{}
}
\description{
server function
}
