% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_classification.R
\name{module_classification_UI}
\alias{module_classification_UI}
\title{classification}
\usage{
module_classification_UI(id)
}
\arguments{
\item{id}{shiny id}

\item{label}{fileInput label}
}
\description{
A shiny Module qui permet de classer des trajectoires
}
