% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{module_data_UI}
\alias{module_data_UI}
\title{type de file, select of ind and data}
\usage{
module_data_UI(id)
}
\arguments{
\item{id}{shiny id}

\item{label}{fileInput label}
}
\description{
A shiny Module that imports data, select them and export as output: data.for.seq, data.comp
}
