% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_select_and_plot.R
\name{module_select_and_plot_UI}
\alias{module_select_and_plot_UI}
\title{select_and_plot}
\usage{
module_select_and_plot_UI(id)
}
\arguments{
\item{id}{shiny id}

\item{label}{fileInput label}
}
\description{
A shiny Module qui permet de sélectionner des données et de représenter les trajectoires
}
