% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tabdes2.R
\name{module_tabdes2}
\alias{module_tabdes2}
\title{server function}
\usage{
module_tabdes2(input, output, session, data)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data}{}
}
\description{
server function
}
