% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tabdes.R
\name{module_tabdes_UI}
\alias{module_tabdes_UI}
\title{tabdes}
\usage{
module_tabdes_UI(id)
}
\arguments{
\item{id}{shiny id}

\item{label}{fileInput label}
}
\description{
A shiny Module qui permet de tabdes
}
