% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdef_modgap.R
\name{seqdef_modgap}
\alias{seqdef_modgap}
\title{La fonction "seqdef_modgap"}
\usage{
seqdef_modgap(
  data,
  var = NULL,
  informat = "STS",
  stsep = NULL,
  alphabet = NULL,
  states = NULL,
  id = NULL,
  weights = NULL,
  start = 1,
  left = NA,
  right = "DEL",
  gaps = "gap",
  missing = NA,
  void = "\%",
  nr = "*",
  cnames = NULL,
  xtstep = 1,
  tick.last = FALSE,
  cpal = NULL,
  missing.color = "darkgrey",
  labels = NULL,
  minimal.gap = 3,
  regle.pour.faux.gap = "before",
  ...
)
}
\arguments{
\item{minimal.gap}{numeric. nombre minimum d'états manquant consécutifs pouvant être considérés comme un "vrai" gap (une sortie du RSA)}

\item{regle.pour.faux.gap}{Choix entre: c("before", "after"). Si "before" ("after"), les faux gap sont recodés avec l'état précédent (suivant) non manquant.}
}
\value{
un vecteur de facteurs res
}
\description{
...

seqdef_modgap: permet de définir un nombre minimal de gaps consécutif pour êtres considérés comme des vrais gaps. Prend en compte tous les paramètres de TraMineR::seqdef() avec deux argumets supplémentaires. 
Attention: l'argument gap ne doit pas être nul. 
Voir les modifs dans seqprep.modgap() (file: seqprep_modgap.R) et dans TraMineR.trunc.modgap() (file: TraMineR_trunc_modgap.R)
}
\author{
Elie
}
