% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction tableau_ligne.R
\name{tableau_ligne}
\alias{tableau_ligne}
\title{Title}
\usage{
tableau_ligne(data, var_grp, var)
}
\arguments{
\item{data}{data frame}

\item{var_grp}{variable contenant la classification}

\item{var}{variable étudiée}
}
\value{
un data frame mis en forme
}
\description{
Title
}
