% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_taille_graph_flux.R
\name{taille_graph_flux}
\alias{taille_graph_flux}
\title{Fonction Taille_graph_flux}
\usage{
taille_graph_flux(nb_grp)
}
\arguments{
\item{nb_grp}{: un entier}
}
\value{
une matrice donnant les positions et l'ordre des graphiques pour la fonction  marrangeGrob pour afficher les graphiques de flux pour chaque groupe
}
\description{
Fonction Taille_graph_flux
}
