% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction vect sous seq.R
\name{vect.sous.seq}
\alias{vect.sous.seq}
\title{Title}
\usage{
vect.sous.seq(data)
}
\arguments{
\item{data}{un data frame dans lequel une colonne correspond à un état et une ligne à une succession d'état. Le nombre d'états successifs est compris
entre 2 et 3.}
}
\value{
un vecteur mise en forme pour être utilisé dans la fonction subsequence comme valeur pour l'argument str.subseq
}
\description{
Title
}
