% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancer_application.R
\name{ViCaTraj}
\alias{ViCaTraj}
\title{Fonction de lancement de l'appliation}
\usage{
ViCaTraj(navigateur = TRUE, ...)
}
\arguments{
\item{navigateur}{lancer dans le navigateur web par défaut?}
}
\description{
Fonction de lancement de l'appliation
}
