% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML_DATA.R
\name{XML_DATA}
\alias{XML_DATA}
\title{La fonction ""seqdef_modgap""XML_DATA"}
\usage{
XML_DATA(emplacement.dossier, ID.var)
}
\arguments{
\item{minimal.gap}{numeric. nombre minimum d'états manquant consécutifs pouvant être considérés comme un "vrai" gap (une sortie du RSA)}

\item{regle.pour.faux.gap}{Choix entre: c("before", "after"). Si "before" ("after"), les faux gap sont recodés avec l'état précédent (suivant) non manquant.}
}
\value{
une liste de data.frame
}
\description{
...

l'objectif de cette fonction est de charger les données présentes dans de multiples objets xml
}
\author{
Elie
}
