% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML_TO_CSV_OLD.R
\name{XML_TO_CSV}
\alias{XML_TO_CSV}
\title{La fonction "XML_TO_CSV"}
\usage{

  XML_TO_CSV(chemin.xml = "rsa.RSABEM.00122018.A0110442.xml/rsa.RSABEM.00122018.A0110442.xml",
  chemin.destination = "RSABEM_201812.csv", return.df = FALSE,
  save.csv = TRUE)
}
\value{
un df et un csv enregistré
}
\description{
...

l'objectif de cette fonction est de charger les données présentes dans de multiples objets xml
}
\author{
Elie
}
