% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction graph flux ggplot couleur.R
\name{graph_flux_grp}
\alias{graph_flux_grp}
\title{Fonction pour faire des graphiques de flux pour un groupe d'individus donné}
\usage{
graph_flux_grp(data, seq_data, col_periode, var_grp, label_grp)
}
\arguments{
\item{data}{}

\item{seq_data}{}

\item{col_periode}{}

\item{var_grp}{}

\item{label_grp}{}
}
\description{
Fonction pour faire des graphiques de flux pour un groupe d'individus donné
}
